move deserialization function
